requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-CORS==4.0.0
mysqlclient==2.2.0
python-dotenv==1.0.0
Flask-RESTful==0.3.10
requests==2.31.0
gunicorn==21.2.0
pymysql==1.1.0
Werkzeug==2.3.7
email-validator==2.1.0
bcrypt==4.0.1
itsdangerous==2.1.2 